+++ KeeAnywhere/Configuration/ConfigurationService.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using KeeAnywhere.StorageProviders;
-using KeePass.App.Configuration;
 using KeePass.Plugins;
 using KeePassLib.Utility;
 using Newtonsoft.Json;
@@ -155,7 +154,7 @@ namespace KeeAnywhere.Configuration
 
             if (string.IsNullOrEmpty(this.PluginConfiguration.BackupToLocalFolder))
             {
-                this.PluginConfiguration.BackupToLocalFolder = Path.Combine(AppConfigSerializer.LocalAppDataDirectory,
+                this.PluginConfiguration.BackupToLocalFolder = Path.Combine(ConfigurationInfo.SettingsDirectory,
                     "KeeAnywhereBackups");
             }